Mantis#1314

+++ far/datetime.cpp
@@ -637,10 +637,10 @@ void StrToDateTime(const wchar_t *CDate, const wchar_t *CTime, FILETIME &ft, int
 	if (bRelative)
 	{
 		ULARGE_INTEGER time;
-		time.QuadPart  = st.wSecond * 10000000;
-		time.QuadPart += st.wMinute * 10000000 * 60;
-		time.QuadPart += st.wHour   * 10000000 * 60 * 60;
-		time.QuadPart += st.wDay    * 10000000 * 60 * 60 * 24;
+		time.QuadPart  = (UINT64)st.wSecond * 10000000ull;
+		time.QuadPart += (UINT64)st.wMinute * 10000000ull * 60ull;
+		time.QuadPart += (UINT64)st.wHour   * 10000000ull * 60ull * 60ull;
+		time.QuadPart += (UINT64)st.wDay    * 10000000ull * 60ull * 60ull * 24ull;
 		ft.dwLowDateTime  = time.u.LowPart;
 		ft.dwHighDateTime = time.u.HighPart;
 	}
@@ -782,13 +782,13 @@ void ConvertDate(const FILETIME &ft,string &strDateText, string &strTimeText,int
 void ConvertRelativeDate(const FILETIME &ft,string &strDaysText,string &strTimeText)
 {
 	ULARGE_INTEGER time={ft.dwLowDateTime,ft.dwHighDateTime};
-	WORD d = (WORD)(time.QuadPart / (10000000ll * 60 * 60 * 24));
-	time.QuadPart = time.QuadPart - ((UINT64)d * 10000000 * 60 * 60 * 24);
-	WORD h = (WORD)(time.QuadPart / (10000000ll * 60 * 60));
-	time.QuadPart = time.QuadPart - ((UINT64)h * 10000000 * 60 * 60);
-	WORD m = (WORD)(time.QuadPart / (10000000 * 60));
-	time.QuadPart = time.QuadPart - ((UINT64)m * 10000000 * 60);
-	WORD s = (WORD)(time.QuadPart / 10000000);
+	WORD d = (WORD)(time.QuadPart / (10000000ull * 60ull * 60ull * 24ull));
+	time.QuadPart = time.QuadPart - ((UINT64)d * 10000000ull * 60ull * 60ull * 24ull);
+	WORD h = (WORD)(time.QuadPart / (10000000ull * 60ull * 60ull));
+	time.QuadPart = time.QuadPart - ((UINT64)h * 10000000ull * 60ull * 60ull);
+	WORD m = (WORD)(time.QuadPart / (10000000ull * 60));
+	time.QuadPart = time.QuadPart - ((UINT64)m * 10000000ull * 60ull);
+	WORD s = (WORD)(time.QuadPart / 10000000ull);
 	strDaysText.Format(L"%u",d);
 	strTimeText.Format(L"%02d%c%02d%c%02d", h, GetTimeSeparator(), m, GetTimeSeparator(), s);
 }
